test: import minimax directly
