+++ src/browser/common.js
@@ -1,14 +1,5 @@
 // Shared utility functions
 
-/**
- * Checks if a value is a non-null object (but not an array).
- * @param {*} val
- * @returns {boolean}
- */
-function isNonNullNonArray(val) {
-  return val !== null && !Array.isArray(val);
-}
+import { isObject } from '../utils/objectUtils.js';
 
-export function isObject(val) {
-  return isNonNullNonArray(val) && typeof val === 'object';
-}
+export { isObject };
