+++ src/toys/2025-03-29/setTemporary.js
@@ -1,5 +1,4 @@
-import { isObject } from '../../browser/common.js';
-import { deepMerge } from '../../browser/data.js';
+import { mergeIntoStateProperty } from '../../utils/stateMerge.js';
 
 /**
  * Parses input as JSON, deep merges it into the 'temporary' object obtained via env.getData(),
@@ -9,39 +8,5 @@ import { deepMerge } from '../../browser/data.js';
  * @returns {string} A confirmation message or an error message.
  */
 export function setTemporary(input, env) {
-  let inputJson;
-  try {
-    inputJson = JSON.parse(input);
-  } catch (parseError) {
-    return `Error: Invalid JSON input. ${parseError.message}`;
-  }
-  return processSetTemporary(inputJson, env);
-}
-
-function processSetTemporary(inputJson, env) {
-  if (!isObject(inputJson)) {
-    return 'Error: Input JSON must be a plain object.';
-  }
-  return handleValidTemporaryInput(inputJson, env);
-}
-
-function handleValidTemporaryInput(inputJson, env) {
-  const getData = env.get('getData');
-  const setData = env.get('setData');
-  try {
-    return mergeTemporaryData(getData, setData, inputJson);
-  } catch (error) {
-    return `Error updating temporary data: ${error.message}`;
-  }
-}
-
-function mergeTemporaryData(getData, setData, inputJson) {
-  const currentData = getData();
-  const newData = JSON.parse(JSON.stringify(currentData));
-  if (!isObject(newData.temporary)) {
-    newData.temporary = {};
-  }
-  newData.temporary = deepMerge(newData.temporary, inputJson);
-  setData(newData);
-  return `Success: Temporary data deep merged.`;
+  return mergeIntoStateProperty('temporary', input, env);
 }
