+++ src/toys/2025-05-08/battleshipSolitaireFleet.js
@@ -11,6 +11,8 @@
  *   • No diagonal placement; honours optional noTouching flag.
  */
 
+import { parseJsonOrDefault } from '../../utils/jsonUtils.js';
+
 // ────────────────────── Helper utilities ────────────────────── //
 
 /** Fisher‑Yates shuffle (in‑place) using env RNG */
@@ -318,13 +320,6 @@ function ensureShipsArray(cfg) {
 
 // ─────────────────────────── Public toy ─────────────────────────── //
 
-function safeJsonParse(input) {
-  try {
-    return JSON.parse(input);
-  } catch {
-    return { width: 10, height: 10, ships: [] };
-  }
-}
 
 function convertShipsToArray(cfg) {
   if (typeof cfg.ships === 'string') {
@@ -348,7 +343,7 @@ function parseDimensions(cfg) {
 }
 
 function parseConfig(input) {
-  const cfg = safeJsonParse(input);
+  const cfg = parseJsonOrDefault(input, { width: 10, height: 10, ships: [] });
   convertShipsToArray(cfg);
   parseDimensions(cfg);
   ensureShipsArray(cfg);
