+++ test/toys/2025-04-06/minimax.earlyreturn.mutant.test.js
@@ -1,22 +1,12 @@
-import fs from 'fs';
-import path from 'path';
 import { describe, test, expect } from '@jest/globals';
-
-async function loadMinimax() {
-  const filePath = path.join(process.cwd(), 'src/toys/2025-04-06/ticTacToe.js');
-  let src = fs.readFileSync(filePath, 'utf8');
-  src += '\nexport { minimax };';
-  const mod = await import(`data:text/javascript,${encodeURIComponent(src)}`);
-  return mod.minimax;
-}
+import { minimax } from '../../../src/toys/2025-04-06/ticTacToe.js';
 
 describe('minimax early return', () => {
-  test('returns terminal score when player has already won', async () => {
-    const minimax = await loadMinimax();
+  test('returns terminal score when player has already won', () => {
     const board = [
       ['X', 'X', 'X'],
       [null, null, null],
-      [null, null, null]
+      [null, null, null],
     ];
     const params = { board, player: 'X', moves: [] };
     const score = minimax(0, true, params);
